pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["sey323 <sey323_nm@yahoo.co.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
line-bot-sdk = "^2.4.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
